requirements.txt
asgiref==3.8.1
Django==5.0.6
image==1.5.33
mysql==0.0.3
mysqlclient==2.2.4
ntchat==0.1.17
pillow==10.3.0
pyee==11.1.0
PyMySQL==1.1.0
PyYAML==6.0.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
watchdog==4.0.0
xcgui==0.1.7
